[CHANGE] - Added remove record props, Added validation of user id in delete income and expense, Added validation of user id on update record

+++ src/records/controllers/records.controller.ts
@@ -99,13 +99,19 @@ export class RecordsController {
   }
 
   @Delete('/expenses')
-  removeExpense(@Body() payload: DeleteRecordDto) {
-    return this.recordsService.removeRecord(payload);
+  removeExpense(@Body() payload: DeleteRecordDto, @Request() req) {
+    const userId = req.user.sub;
+    return this.recordsService.removeRecord({ payload, userId });
   }
 
   @Delete('/incomes')
-  removeIncome(@Body() payload: DeleteRecordDto) {
-    return this.recordsService.removeRecord(payload, true);
+  removeIncome(@Body() payload: DeleteRecordDto, @Request() req) {
+    const userId = req.user.sub;
+    return this.recordsService.removeRecord({
+      payload,
+      userId,
+      isIncome: true,
+    });
   }
 
   @Post('/expenses/multiple')
